move qat activation quantizers to trainable infra
