+++ model_compression_toolkit/trainable_infrastructure/keras/activation_quantizers/ste/uniform_ste.py
@@ -12,61 +12,47 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-import numpy as np
 import tensorflow as tf
 from tensorflow.python.framework.tensor_shape import TensorShape
-from model_compression_toolkit.constants import RANGE_MIN, RANGE_MAX
-from model_compression_toolkit.qat.keras.quantizer.base_keras_qat_weight_quantizer import \
-    BaseKerasQATWeightTrainableQuantizer
 from model_compression_toolkit.trainable_infrastructure.common.constants import FQ_MIN, FQ_MAX
 from model_compression_toolkit.trainable_infrastructure import KerasTrainableQuantizationWrapper
 from model_compression_toolkit.trainable_infrastructure import TrainingMethod
 
 from mct_quantizers import mark_quantizer, QuantizationMethod, QuantizationTarget
-from mct_quantizers.keras.quantizers import BaseKerasInferableQuantizer, WeightsUniformInferableQuantizer
+from mct_quantizers.keras.quantizers import BaseKerasInferableQuantizer, ActivationUniformInferableQuantizer
 
 from model_compression_toolkit.qat.keras.quantizer.quant_utils import adjust_range_to_include_zero
 from model_compression_toolkit.core.common.quantization.quantizers.quantizers_helpers import fix_range_to_include_zero
+from model_compression_toolkit import constants as C
 
-from model_compression_toolkit.trainable_infrastructure import TrainableQuantizerWeightsConfig
+from model_compression_toolkit.trainable_infrastructure import TrainableQuantizerActivationConfig
 from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import VariableGroup
+from model_compression_toolkit.trainable_infrastructure.keras.activation_quantizers import BaseKerasActivationTrainableQuantizer
 
 
-@mark_quantizer(quantization_target=QuantizationTarget.Weights,
+@mark_quantizer(quantization_target=QuantizationTarget.Activation,
                 quantization_method=[QuantizationMethod.UNIFORM],
                 identifier=TrainingMethod.STE)
-class STEUniformWeightQATQuantizer(BaseKerasQATWeightTrainableQuantizer):
+class STEUniformActivationTrainableQuantizer(BaseKerasActivationTrainableQuantizer):
     """
-    Trainable constrained quantizer to quantize a layer inputs.
+    Trainable constrained quantizer to quantize a layer outputs.
     """
 
-    def __init__(self, quantization_config: TrainableQuantizerWeightsConfig):
+    def __init__(self, quantization_config: TrainableQuantizerActivationConfig, freeze_quant_params: bool = False):
         """
-        Initialize a TrainableWeightQuantizer object with parameters to use
+        Initialize a STEUniformActivationTrainableQuantizer object with parameters to use
         for the quantization.
 
         Args:
-            quantization_config: a trainable quantizer config class with attributes for the quantization.
+            quantization_config: trainable quantizer config class
+            freeze_quant_params: whether to freeze learnable quantization parameters. This is unused here, since there is not any quantizaiton params that are learned.
 
         """
-        super().__init__(quantization_config)
-        self.max_values = np.array(quantization_config.weights_quantization_params[RANGE_MAX])
-        self.min_values = np.array(quantization_config.weights_quantization_params[RANGE_MIN])
-        self.num_bits = self.quantization_config.weights_n_bits
-        self.per_channel = self.quantization_config.weights_per_channel_threshold
-        self.channel_axis = self.quantization_config.weights_channels_axis
-        self.min_max_shape = np.asarray(self.max_values).shape
-        self.max = np.reshape(self.max_values, [-1]) if self.per_channel else float(self.max_values)
-        self.min = np.reshape(self.min_values, [-1]) if self.per_channel else float(self.min_values)
+        super().__init__(quantization_config, freeze_quant_params)
 
-        if self.per_channel and self.channel_axis not in [-1, len(self.min_max_shape) - 1]:
-            # Tensorflow's fake_quant_with_min_max_vars_per_channel only works on last axis, so
-            # need to move the quantization axis to the last axis
-            self.perm_vec = list(np.arange(len(self.min_max_shape)))
-            self.perm_vec[self.channel_axis] = len(self.min_max_shape) - 1
-            self.perm_vec[len(self.min_max_shape) - 1] = self.channel_axis
-        else:
-            self.perm_vec = None
+        self.num_bits = quantization_config.activation_n_bits
+        self.min_range = quantization_config.activation_quantization_params[C.RANGE_MIN]
+        self.max_range = quantization_config.activation_quantization_params[C.RANGE_MAX]
 
     def initialize_quantization(self,
                                 tensor_shape: TensorShape,
@@ -82,23 +68,24 @@ class STEUniformWeightQATQuantizer(BaseKerasQATWeightTrainableQuantizer):
         """
         fq_min = layer.add_weight(
             name + FQ_MIN,
-            shape=len(self.min) if self.per_channel else (),
+            shape=(),
             initializer=tf.keras.initializers.Constant(-1.0),
             trainable=False)
-        fq_min.assign(self.min)
+        fq_min.assign(self.min_range)
 
         fq_max = layer.add_weight(
             name + FQ_MAX,
-            shape=len(self.max) if self.per_channel else (),
+            shape=(),
             initializer=tf.keras.initializers.Constant(1.0),
             trainable=False)
-        fq_max.assign(self.max)
+        fq_max.assign(self.max_range)
 
         # save the quantizer added parameters for later calculations
         self.add_quantizer_variable(FQ_MIN, fq_min, VariableGroup.QPARAMS)
         self.add_quantizer_variable(FQ_MAX, fq_max, VariableGroup.QPARAMS)
 
-    def __call__(self, inputs: tf.Tensor,
+    def __call__(self,
+                 inputs: tf.Tensor,
                  training: bool):
         """
         Quantize a tensor.
@@ -113,18 +100,8 @@ class STEUniformWeightQATQuantizer(BaseKerasQATWeightTrainableQuantizer):
         _min = self.get_quantizer_variable(FQ_MIN)
         _max = self.get_quantizer_variable(FQ_MAX)
         _min, _max = adjust_range_to_include_zero(_min, _max, self.num_bits)
-
-        if self.per_channel:
-            if self.perm_vec:
-                inputs = tf.transpose(inputs, perm=self.perm_vec)
-
-            q_tensor = tf.quantization.fake_quant_with_min_max_vars_per_channel(inputs, _min, _max,
-                                                                                num_bits=self.num_bits)
-            if self.perm_vec:
-                q_tensor = tf.transpose(q_tensor, perm=self.perm_vec)
-        else:
-            q_tensor = tf.quantization.fake_quant_with_min_max_vars(inputs, _min, _max,
-                                                                    num_bits=self.num_bits)
+        q_tensor = tf.quantization.fake_quant_with_min_max_vars(inputs, _min, _max,
+                                                                num_bits=self.num_bits)
 
         return q_tensor
 
@@ -138,11 +115,8 @@ class STEUniformWeightQATQuantizer(BaseKerasQATWeightTrainableQuantizer):
         min_range, max_range = fix_range_to_include_zero(self.get_quantizer_variable(FQ_MIN).numpy(),
                                                          self.get_quantizer_variable(FQ_MAX).numpy(),
                                                          self.num_bits)
-        return WeightsUniformInferableQuantizer(num_bits=self.num_bits,
-                                                min_range=list(min_range.flatten()),
-                                                max_range=list(max_range.flatten()),
-                                                per_channel=self.per_channel,
-                                                channel_axis=self.channel_axis,
-                                                input_rank=len(self.min_max_shape))
-
-
+        return ActivationUniformInferableQuantizer(num_bits=self.num_bits,
+                                                   # In activation quantization is per-tensor only - thus we pass
+                                                   # the min/max as lists with a len of 1
+                                                   min_range=[min_range],
+                                                   max_range=[max_range])
