+++ tests/keras_tests/feature_networks_tests/feature_networks/qat/qat_test.py
@@ -23,10 +23,13 @@ from mct_quantizers import QuantizationTarget, KerasActivationQuantizationHolder
 from mct_quantizers.common.base_inferable_quantizer import QuantizerID
 from mct_quantizers.common.get_all_subclasses import get_all_subclasses
 from mct_quantizers.keras.quantizers import BaseKerasInferableQuantizer
-from model_compression_toolkit.qat.keras.quantizer.base_keras_qat_quantizer import BaseKerasQATTrainableQuantizer
+from model_compression_toolkit.qat.keras.quantizer.base_keras_qat_weight_quantizer import \
+    BaseKerasQATWeightTrainableQuantizer
 from model_compression_toolkit.trainable_infrastructure import TrainingMethod, KerasTrainableQuantizationWrapper, \
     BaseKerasTrainableQuantizer
 from model_compression_toolkit.trainable_infrastructure.common.base_trainable_quantizer import BaseTrainableQuantizer
+from model_compression_toolkit.trainable_infrastructure.keras.activation_quantizers import \
+    BaseKerasActivationTrainableQuantizer
 from model_compression_toolkit.trainable_infrastructure.keras.load_model import \
     keras_load_quantized_model
 from tests.keras_tests.feature_networks_tests.base_keras_feature_test import BaseKerasFeatureNetworkTest
@@ -147,7 +150,7 @@ class QuantizationAwareTrainingQuantizerHolderTest(QuantizationAwareTrainingTest
         holder_layers = get_layers_from_model_by_type(quantized_model, KerasActivationQuantizationHolder)
         self.unit_test.assertTrue(len(holder_layers)==2)
         for holder_layer in holder_layers:
-            self.unit_test.assertTrue(holder_layer.get_config()['activation_holder_quantizer']['class_name'] == 'STEActivationQATQuantizer')
+            self.unit_test.assertTrue(holder_layer.get_config()['activation_holder_quantizer']['class_name'] == 'STESymmetricActivationTrainableQuantizer')
             self.unit_test.assertTrue(holder_layer.get_config()['activation_holder_quantizer']['config']['quantization_config']['activation_n_bits'] == 4)
 
             # Assert weights of quantizer were added to the quantization holder layer
@@ -233,7 +236,9 @@ class QATWrappersTest(BaseKerasFeatureNetworkTest):
             self.unit_test.assertTrue(np.isclose(np.linalg.norm(out_qat_finalize_model - out_ptq_model) / np.linalg.norm(out_ptq_model), 0, atol=1e-6))
 
     def compare(self, qat_model, finalize=False, input_x=None, quantization_info=None):
-        all_trainable_quantizers = get_all_subclasses(BaseKerasQATTrainableQuantizer)
+        all_trainable_quantizers = get_all_subclasses(BaseKerasQATWeightTrainableQuantizer)
+        all_trainable_quantizers = all_trainable_quantizers.union(get_all_subclasses(BaseKerasActivationTrainableQuantizer))
+
         all_inferable_quantizers = get_all_subclasses(BaseKerasInferableQuantizer)
         for layer in qat_model.layers:
             if isinstance(layer, KerasActivationQuantizationHolder):
